move age validation tests to shared validation spec
